+++ R/ggMarginal.R
@@ -27,6 +27,14 @@
 #' the x axis.
 #' @param yparams List of extra parameters to use only for the marginal plot along
 #' the y axis.
+#' @param groupColour Do you want the colour mapping from your scatter plot to 
+#' to used to define the colours of the marginal density plots? If \code{TRUE}, 
+#' the variable mapped to \code{colour} in your scatter plot must be a character 
+#' or factor variable. See examples.
+#' @param groupFill Do you want the colour mapping from your scatter plot to 
+#' to used to define the fills of the marginal density plots? If \code{TRUE}, 
+#' the variable mapped to \code{colour} in your scatter plot must be a character 
+#' or factor variable. See examples.
 #' @return An object of class \code{ggExtraPlot}. This object can be printed to show the
 #' plots or saved using any of the typical image-saving functions (for example, using
 #' \code{png()} or \code{pdf()}).
@@ -77,11 +85,23 @@
 #'
 #' p4 <- p3 + ggplot2::scale_x_continuous(limits = c(2, 6)) + ggplot2::theme_bw(50)
 #' ggMarginal(p4)
+#' 
+#' # Using groupColour and groupFill
+#' # In order to use either of these arguments, we must map colour in the scatter 
+#' # plot to a factor or character variable (e.g., the vs variable below)
+#' p <- ggplot2::ggplot(mtcars, ggplot2::aes(x = wt, y = drat, colour = factor(vs))) +
+#'      ggplot2::geom_point()
+#' # Now we can color or fill the marginal density plots using the same groupings
+#' # as we used in the scatter plot
+#' ggMarginal(p, groupColour = TRUE)
+#' ggMarginal(p, groupColour = TRUE, groupFill = TRUE)
+#' 
 #' @seealso \href{http://daattali.com/shiny/ggExtra-ggMarginal-demo/}{Demo Shiny app}
 #' @export
 ggMarginal <- function(p, data, x, y, type = c("density", "histogram", "boxplot", "violin"),
                        margins = c("both", "x", "y"), size = 5,
-                       ..., xparams = list(), yparams = list()) {
+                       ..., xparams = list(), yparams = list(),
+                       groupColour = FALSE, groupFill = FALSE) {
 
   # Figure out all the default parameters.
   type <- match.arg(type)
@@ -89,14 +109,9 @@ ggMarginal <- function(p, data, x, y, type = c("density", "histogram", "boxplot"
 
   # Fill in param defaults and consolidate params into single list (prmL).
   prmL <- toParamList(exPrm = list(...), xPrm = xparams, yPrm = yparams)
-
-  # After ggplot2 v1.0.1, layers became strict about parameters
-  if (type == "density") {
-    prmL[['exPrm']][['fill']] <- NULL
-  }
-
-  # Create one version of the scat plot (scatP), based on values of p, data, x,
-  # and y...also remove all margin around plot so that it's easier to position
+  
+  # Create one version of the scat plot (scatP), based on values of p, data, x, 
+  # and y...also remove all margin around plot so that it's easier to position 
   # the density plots beside the main plot
   scatP <- reconcileScatPlot(p = p, data = data, x = x, y = y) +
     ggplot2::theme(plot.margin = grid::unit(c(0, 0, .25, .25), "cm"))
@@ -123,16 +138,18 @@ ggMarginal <- function(p, data, x, y, type = c("density", "histogram", "boxplot"
 
   # If margins = x or 'both' (x and y), then you have to create top plot
   # Top plot = horizontal margin plot, which corresponds to x marg
-  if (margins != "y") {
-    top <- genFinalMargPlot(marg = "x", type = type, scatPbuilt = scatPbuilt,
-                            prmL = prmL)
+  if (margins != "y") { 
+    top <- genFinalMargPlot(marg = "x", type = type, scatPbuilt = scatPbuilt, 
+                            prmL = prmL, groupColour = groupColour, 
+                            groupFill = groupFill)
   }
 
   # If margins = y or 'both' (x and y), then you have to create right plot.
   # (right plot = vertical margin plot, which corresponds to y marg)
-  if (margins != "x") {
-    right <- genFinalMargPlot(marg = "y", type = type, scatPbuilt = scatPbuilt,
-                              prmL = prmL)
+  if (margins != "x") { 
+    right <- genFinalMargPlot(marg = "y", type = type, scatPbuilt = scatPbuilt, 
+                              prmL = prmL, groupColour = groupColour, 
+                              groupFill = groupFill)
   }
 
   # Now add the marginal plots to the scatter plot